- neon stencil2D for r = 1

+++ benchmarks_new/stencil_2D/stencil_2D_main.h
@@ -8,7 +8,7 @@
 #include "../../locutils_new/perf_wrapper.h"
 
 
-typedef double base_type;
+typedef float base_type;
 
 #include "stencil_2D.h"
 #include "../../locutils_new/timers.h"
